TryCatch added
tryCatch is added to the breakpoint calculation, the rest of the code is executed on the condition that no error occured during calculation of breakpoints. If the error did occur, all indicators are set to NA.
This is done to avoid the whole funciton terminating when an error occurs. The error: Error in my.RSS.table[as.character(i), 3:4] <- c(pot.index[opt], break.RSS[opt]) :
    replacement has length zero
occurs in rare cases. Mostly when h is very small and when there are many NAs, probably due to singularities in the regression. However, it is unpredictable when this occurs and on large datasets there always seem to be a few pixels that are susceptible to this error.